# tools/api/mmc_atlas_api_client/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: MMC Atlas API Helper
  title: MMC Atlas API Helper
  version: "1.0"
servers:
- url: //localhost:8080/api
paths:
...
  /api/project/{ProjectId}/databaseUsers:
    get:
      description: Get all database users
      operationId: GetAllDatabaseUser
      parameters:
      - description: Project ID
        explode: false
        in: path
        name: ProjectId
        required: true
        schema:
          default: <projectID>
          type: string
        style: simple
      - description: Public Key
        explode: false
        in: header
        name: x-mongo-publickey
        required: true
        schema:
          default: <publicKey>
          type: string
        style: simple
      - description: Private Key
        explode: false
        in: header
        name: x-mongo-privatekey
        required: true
        schema:
          default: <privateKey>
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/database_user.Model'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Internal Server Error
      summary: Get all database users
      tags:
      - Database User
    post:
      description: Create a new database user with the specified name and email
      operationId: CreateDatabaseUser
      parameters:
      - description: Project ID
        explode: false
        in: path
        name: ProjectId
        required: true
        schema:
          default: <projectID>
          type: string
        style: simple
      - description: Public Key
        explode: false
        in: header
        name: x-mongo-publickey
        required: true
        schema:
          default: <publicKey>
          type: string
        style: simple
      - description: Private Key
        explode: false
        in: header
        name: x-mongo-privatekey
        required: true
        schema:
          default: <privateKey>
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/database_user.InputModel'
        description: body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/database_user.Model'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Internal Server Error
      summary: Create a database user
      tags:
      - Database User
      x-codegen-request-body-name: InputModel
  /api/project/{ProjectId}/databaseUsers/{Username}:
    delete:
      description: Delete a database user with the specified ID
      operationId: DeleteDatabaseUser
      parameters:
      - description: Project ID
        explode: false
        in: path
        name: ProjectId
        required: true
        schema:
          default: <projectID>
          type: string
        style: simple
      - description: Username
        explode: false
        in: path
        name: Username
        required: true
        schema:
          default: testUser
          type: string
        style: simple
      - description: Public Key
        explode: false
        in: header
        name: x-mongo-publickey
        required: true
        schema:
          default: <publicKey>
          type: string
        style: simple
      - description: Private Key
        explode: false
        in: header
        name: x-mongo-privatekey
        required: true
        schema:
          default: <privateKey>
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Internal Server Error
      summary: Delete a database user
      tags:
      - Database User
    get:
      description: Get a database user with the specified ID
      operationId: GetDatabaseUser
      parameters:
      - description: Project ID
        explode: false
        in: path
        name: ProjectId
        required: true
        schema:
          default: <projectID>
          type: string
        style: simple
      - description: Username
        explode: false
        in: path
        name: Username
        required: true
        schema:
          default: testUser
          type: string
        style: simple
      - description: Public Key
        explode: false
        in: header
        name: x-mongo-publickey
        required: true
        schema:
          default: <publicKey>
          type: string
        style: simple
      - description: Private Key
        explode: false
        in: header
        name: x-mongo-privatekey
        required: true
        schema:
          default: <privateKey>
          type: string
        style: simple
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Internal Server Error
      summary: Get a database user
      tags:
      - Database User
components:
  schemas:
...
    database_user.InputModel:
      properties:
        password:
          example: testPass
          type: string
        projectId:
          example: <projectID>
          type: string
        username:
          example: testUser
          type: string
      type: object
    database_user.LabelDefinition:
      example:
        value: value
        key: key
      properties:
        key:
          type: string
        value:
          type: string
      type: object
    database_user.Model:
      example:
        password: password
        awsIAMType: awsIAMType
        databaseName: databaseName
        deleteAfterDate: deleteAfterDate
        profile: profile
        roles:
        - databaseName: databaseName
          roleName: roleName
          collectionName: collectionName
        - databaseName: databaseName
          roleName: roleName
          collectionName: collectionName
        scopes:
        - name: name
          type: type
        - name: name
          type: type
        x509Type: x509Type
        ldapAuthType: ldapAuthType
        projectId: projectId
        labels:
        - value: value
          key: key
        - value: value
          key: key
        username: username
      properties:
        awsIAMType:
          type: string
        databaseName:
          type: string
        deleteAfterDate:
          type: string
        labels:
          items:
            $ref: '#/components/schemas/database_user.LabelDefinition'
          type: array
        ldapAuthType:
          type: string
        password:
          type: string
        profile:
          type: string
        projectId:
          type: string
        roles:
          items:
            $ref: '#/components/schemas/database_user.RoleDefinition'
          type: array
        scopes:
          items:
            $ref: '#/components/schemas/database_user.ScopeDefinition'
          type: array
        username:
          type: string
        x509Type:
          type: string
      type: object
    database_user.RoleDefinition:
      example:
        databaseName: databaseName
        roleName: roleName
        collectionName: collectionName
      properties:
        collectionName:
          type: string
        databaseName:
          type: string
        roleName:
          type: string
      type: object
    database_user.ScopeDefinition:
      example:
        name: name
        type: type
      properties:
        name:
          type: string
        type:
          type: string
      type: object
    AtlasResponse:
      example:
        response: "{}"
        errorMessage: errorMessage
        errorCode: errorCode
        status: status
      properties:
        errorCode:
          type: string
        errorMessage:
          type: string
        response:
          type: object
        status:
          type: string
      type: object
  securitySchemes:
    BasicAuth:
      scheme: basic
      type: http
...